+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.1] - 2019-11-25
+
+### Changes
+
+- Make sure version control integration grants write access before trying to save package settings.
+
+### Bug Fixes
+
+- Fixed samples not compiling with Unity 2019.3.
+- Fix package settings repo potentially initializing with a null dictionary.
+
 ## [1.0.0] - 2019-04-03
 
 ### Bug Fixes